+++ Mage/Repository/Observation/ObservationMapFeatureRepository.swift
@@ -55,29 +55,3 @@ class ObservationMapFeatureRepository: MapFeatureRepository, ObservableObject {
         return []
     }
 }
-
-class ObservationMapItemAnnotation: DataSourceAnnotation {
-//    var id: URL?
-//    var coordinate: CLLocationCoordinate2D
-    var mapItem: ObservationMapItem
-    var title: String?
-    var subtitle: String?
-    
-    override var dataSource: any DataSourceDefinition {
-        get {
-            DataSources.observation
-        }
-        set { }
-    }
-
-    init(mapItem: ObservationMapItem) {
-        self.mapItem = mapItem
-        let itemKey = mapItem.observationLocationId?.absoluteString ?? ""
-        if let point = mapItem.geometry?.centroid() {
-            super.init(coordinate: CLLocationCoordinate2D(latitude: point.y.doubleValue, longitude: point.x.doubleValue), itemKey: itemKey)
-        } else {
-            super.init(coordinate: kCLLocationCoordinate2DInvalid, itemKey: itemKey)
-        }
-        self.id = mapItem.observationLocationId?.absoluteString ?? UUID().uuidString
-    }
-}
